feat: adding env and create interceptor
