+++ .env.example
@@ -1,0 +1,1 @@
+BASE_URL=
