+++ src/types/base-response.ts
@@ -1,0 +1,6 @@
+export interface IBaseResponse {
+  status: boolean
+  code: number
+  messages: string
+  data: any
+}
