+++ src/types/team.ts
@@ -1,0 +1,5 @@
+export interface ICreateTeamRequest {
+  team_name: string
+  icon: string
+  user_id: number
+}
